jsmp.nav

+++ jsmp.nav/src/main/java/jetsennet/jsmp/nav/entity/ChannelEntity.java
@@ -8,7 +8,7 @@ import jetsennet.orm.annotation.Column;
 import jetsennet.orm.annotation.Table;
 import jetsennet.orm.annotation.Id;
 /**
- * 本文件由jetsennet.jsmp.nav.util.BDBFileParse生成，最好不要手动修改！
+ * 
  */
 @Table("NS_CHANNEL")
 public class ChannelEntity implements Serializable
@@ -101,7 +101,7 @@ public class ChannelEntity implements Serializable
 	 * 更新时间
 	 */
 	@Column("UPDATE_TIME")
-	private int updateTime;
+	private long updateTime;
 
 	private static final long serialVersionUID = 1L;
 
@@ -255,12 +255,12 @@ public class ChannelEntity implements Serializable
 		this.regionCode = regionCode;
 	}
 
-	public int getUpdateTime()
+	public long getUpdateTime()
 	{
 		return updateTime;
 	}
 
-	public void setUpdateTime(int updateTime)
+	public void setUpdateTime(long updateTime)
 	{
 		this.updateTime = updateTime;
 	}
