+++ jsmp.nav/src/main/java/jetsennet/jsmp/nav/entity/ProductEntity.java
@@ -7,7 +7,7 @@ import jetsennet.orm.annotation.Column;
 import jetsennet.orm.annotation.Table;
 import jetsennet.orm.annotation.Id;
 /**
- * 本文件由jetsennet.jsmp.nav.util.BDBFileParse生成，最好不要手动修改！
+ * 本文件由jetsennet.jsmp.nav.util.BDBFileParse生成
  */
 @Table("NS_PRODUCT")
 public class ProductEntity implements Serializable
@@ -57,7 +57,7 @@ public class ProductEntity implements Serializable
 	 * 更新时间
 	 */
 	@Column("UPDATE_TIME")
-	private int updateTime;
+	private long updateTime;
 
 	private static final long serialVersionUID = 1L;
 
@@ -141,12 +141,12 @@ public class ProductEntity implements Serializable
 		this.productDesc = productDesc;
 	}
 
-	public int getUpdateTime()
+	public long getUpdateTime()
 	{
 		return updateTime;
 	}
 
-	public void setUpdateTime(int updateTime)
+	public void setUpdateTime(long updateTime)
 	{
 		this.updateTime = updateTime;
 	}
