+++ jsmp.nav/src/main/java/jetsennet/jsmp/nav/entity/ProgramEntity.java
@@ -9,7 +9,7 @@ import jetsennet.orm.annotation.Table;
 import jetsennet.orm.annotation.Id;
 
 /**
- * 本文件由jetsennet.jsmp.nav.util.BDBFileParse生成，最好不要手动修改！
+ * 本文件由jetsennet.jsmp.nav.util.BDBFileParse生成
  */
 @Table("NS_PROGRAM")
 public class ProgramEntity implements Serializable
@@ -97,7 +97,7 @@ public class ProgramEntity implements Serializable
      * 更新时间
      */
     @Column("UPDATE_TIME")
-    private int updateTime;
+    private long updateTime;
     /**
      * 栏目下排序号
      */
@@ -284,12 +284,12 @@ public class ProgramEntity implements Serializable
         this.licensingEnd = licensingEnd;
     }
 
-    public int getUpdateTime()
+    public long getUpdateTime()
     {
         return updateTime;
     }
 
-    public void setUpdateTime(int updateTime)
+    public void setUpdateTime(long updateTime)
     {
         this.updateTime = updateTime;
     }
